Add a parser class to handle all inputs
